Remove QueryableExtensions.Select method which was being used for internal purposes only
