+++ src/EFCore/Internal/EntityFinder.cs
@@ -5,11 +5,11 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
+using System.Reflection;
 using System.Threading;
 using System.Threading.Tasks;
 using JetBrains.Annotations;
 using Microsoft.EntityFrameworkCore.ChangeTracking.Internal;
-using Microsoft.EntityFrameworkCore.Extensions.Internal;
 using Microsoft.EntityFrameworkCore.Metadata;
 using Microsoft.EntityFrameworkCore.Metadata.Internal;
 using Microsoft.EntityFrameworkCore.Storage;
@@ -285,8 +285,30 @@ namespace Microsoft.EntityFrameworkCore.Internal
                 return (IQueryable)((IDbSetCache)context).GetOrAddSet(context.GetDependencies().SetSource, entityType.ClrType);
             }
 
-            return BuildQueryRoot(context, definingEntityType)
-                .Select(entityType.DefiningNavigationName, definingEntityType.ClrType, entityType.ClrType);
+            return BuildQueryRoot(context, definingEntityType, entityType);
+        }
+
+        private static IQueryable BuildQueryRoot(DbContext context, IEntityType definingEntityType, IEntityType entityType)
+        {
+            var queryRoot = BuildQueryRoot(context, definingEntityType);
+
+            return (IQueryable)_selectMethod.MakeGenericMethod(definingEntityType.ClrType, entityType.ClrType)
+                .Invoke(null, new object[] { queryRoot, entityType.DefiningNavigationName });
+        }
+
+        private static readonly MethodInfo _selectMethod
+            = typeof(EntityFinder<TEntity>).GetTypeInfo().GetDeclaredMethods(nameof(Select)).Single(mi => mi.IsGenericMethodDefinition);
+
+        private static IQueryable<TResult> Select<TSource, TResult>(
+            [NotNull] IQueryable<TSource> source, [NotNull] string propertyName)
+            where TResult : class
+            where TSource : class
+        {
+            var parameter = Expression.Parameter(typeof(TSource), "e");
+            return source.Select(
+                Expression.Lambda<Func<TSource, TResult>>(
+                    Expression.MakeMemberAccess(parameter, typeof(TSource).GetAnyProperty(propertyName)),
+                    parameter));
         }
 
         private static BinaryExpression BuildPredicate(
